+++ pdi/docs/First_steps/hello_event.f90
@@ -1,3 +1,7 @@
+! SPDX-FileCopyrightText: 2021-2026 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+!
+! SPDX-License-Identifier: BSD-3-Clause
+
 program hello_event
     use paraconf
     use PDI