+++ CHANGELOG.md
@@ -23,6 +23,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Security
 
 
+## [1.10.1] - 2026-02-04
+
+### Added
+* Added tests for using an installed PDI
+
+
 ## [1.10.0] - 2026-01-31
 
 ### Added